Added larger numbers.

+++ Calculator/Operations/Division.cs
@@ -5,9 +5,9 @@
     {
         public static string Result { get; set; }
 
-        public static string Calculate(string firstNumber, int iterationCount)
+        public static string Calculate(string firstNumber, ulong iterationCount)
         {
-            for (int i = 0; i < iterationCount; i++)
+            for (ulong i = 0; i < iterationCount; i++)
             {
                 Result = OperationSelector.Perform(firstNumber, '-', firstNumber);
             }
