+++ Calculator/Services/BinaryToDecimal.cs
@@ -4,12 +4,12 @@
     public class BinaryToDecimal
     {
         private static string _binaryNumber;
-        public static int DecimalResult { get; set; }
+        public static ulong DecimalResult { get; set; }
 
-        public static int Converter(string binary)
+        public static ulong Converter(string binary)
         {
             _binaryNumber = binary;
-            DecimalResult = Convert.ToInt32(_binaryNumber, 2);
+            DecimalResult = Convert.ToUInt64(_binaryNumber, 2);
             return DecimalResult;
         }
     }
